requirements.txt
asana==5.0.7
openai>=1.51.1
python-dotenv==0.13.0
langchain==0.2.6
langchain-anthropic==0.1.16
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.10
streamlit==1.36.0
requests==2.31.0
flask==3.0.3
pyjwt==2.1.0
pinecone==3.1.0
backoff==2.2.1

